Renamed "PseudoRandomTextGenerator" to "TextTransformers" TextGenerator class renamed to MarkovGenerator TextMunger console tester renamed to Runner; some bits are currently being used from the GUI; will be refactored out. Added empty Transformer factory class file. TODO: build it.
added process button and output window to gui.
Forced and input source for testing
Naive looping through two-assumed rule-sets and application of rules.
An output appears!

Although it seems that no actual rules are selected, but they appear anyway. So that needs to be looked into.
And editors built.
And a smarter looping of rules -- since there should be pre-globals, and post-globals, and basically, however the user wants them ordered.
Only we shouldn't mix the granularity of RuleSets -- so granularity should be a property, and the rule-factories need to respect that.

ah, so much work, but proofs-of-concept are pleasant!!!