shrank the form to fit on a smaller screen by default. TODO: resize form code
added required libraries to project and SVN
Including all of NUnit for unit tests that I'm not even running. but should.

working on properly saving and pre-selecting the Markov rule-type.
Need to think about this, there's too many pieces in play.

Storing originally texts so they can be pre-populated on re-opening the libary selector.
Need to buil populate code, but also need to take into account clearing the SOURCE text -- if desired.
Also, want to swap OUTPUT for SOURCE, if we want to process it AGAIN in some fashion....
For that, we should be storing the SOURCE somewhere, so we can restore it -- as it may have been edited.
Too much for memory, put in a temp file...