working on adding serialization to Texts and libraries Can serialize a text (no save/load tested -- only to JSON string and back) But can't serialize the enclosing library, which is half the point (ie, a library full of texts). Dictionary<string, Text> seems to be the issue. Although Text is serializable, the dictionary is NOT, using this method?
Look into using JavaScriptSerializer or going to XMLSerializer

Test code added in ConsoleRunner.
Extension methods added to TextSourcers
Would have added to GUI, but that meant that TextSources needed to reference GUI, and that was a dis-allowed circular reference.
Might have added some references that are no longer needed as locations changed.....