Added two new Transformers: ShortLines (configurable newlines randomly after words) and Initial spaces (configurably random spaces added to beginning of lines)
The two were then composited to replace the previous implementation of FreeVerse.
So now there are three rules, and some can be used separately...

TODO: a couple of other rules that could go into Free Verse, or something
Like HejinianMemoryAid that adds spaces so that first letter is in corresponding alpha position (eg, a=0 b=1 leading spaces, etc)
chop off leading portions of lines (nice prior to HeijinianMemoryAid)
better implementation of newLines for eecummingsification -- need a word-splitter
or is that separate?