Load RuleSets works from save RuleSets. Also implemented auto-load predefined rulesets into available-list NOT the pre-selected list. Althogh maybe that should be configurable?
TODO: markov generate gets really weird when it encounters a key that is not in the dictionary.
It get the next bit at random, which can lead to a long loop of unknown keys, and totally random gibberish.
Which has a much higher order of entropy than the surrounding text