Implemented ShortLines editor, update InitialSpaces editor. Both will default an empty field on save to 0 (may be blank pending an edit). Both added and launch. Not fully tested to confirm save, etc. but roughly they look good.
Refactored split-to-words from AidToMemory, and implemented Split-to-paragraphs.
Split-to-paragraphs needs work, it's fixed at two-return-chars for split, and should be 2-or-more
TODO: these might make sense as classes, with a granularity, and could be provided as parameters to rules.

Implemented ShuffleParagraph -- as an example of a different shuffle granularity.
It gives the idea of what could happen with different params, perhaps.