Added CaseSensitivity on/off for Markov generator, and exposed in editor.
case-sensitive character n-grams, the former default, now look crap, with more gibberish and mixed-case.
NOT case-senstive character n-grams look good, which is NOT what I would be expecting.

I do note that case-sensitive character n-grams hit the "key-not-found" logic, and call the new InitializeKey() routine.
That was just added.

could that be the source of the problem?!??