FIXED.
The case-sensitivity code was INVERTED, and I somehow managed to not notice, despite stepping through that section of code several times.

It's the NOT case-sensitive n-grams that end up with more keys-not-found, and multiple re-initializations of the key as a result.
Leading to more randomness.

On a word-level, this is some cognitive jumps, but not extra jarring.
On a character-level, words vanish, mixed-case agglomerations appear, and there is little sense of statistical chaining.