Working on XRML issues, getting close.
Added command-line-options parser (NDesk.Options), implemented option for auto-loading a RuleSet.
As a results, discovered a bug and workaround in the CustomControl.

Fixed small bug in RuleSetEditor where it would try to process a null rule.
The root-cause is a click on a blank-line; not even sure why that would fire.
Which means this might not be the correct solution....